Move scheduler code from tp_worker.py to scheduler.py (#1538)
